Move some Communicator methods to communicator.C
These aren't templated or speed-critical; there's really no reason we
need them in a header.
