+++ include/parallel/parallel_implementation.h
@@ -149,168 +149,6 @@ namespace libMesh
 namespace Parallel
 {
 
-inline Communicator::Communicator () :
-#ifdef LIBMESH_HAVE_MPI
-  _communicator(MPI_COMM_SELF),
-#endif
-  _rank(0),
-  _size(1),
-  _send_mode(DEFAULT),
-  used_tag_values(),
-  _I_duped_it(false) {}
-
-inline Communicator::Communicator (const communicator & comm) :
-#ifdef LIBMESH_HAVE_MPI
-  _communicator(MPI_COMM_SELF),
-#endif
-  _rank(0),
-  _size(1),
-  _send_mode(DEFAULT),
-  used_tag_values(),
-  _I_duped_it(false)
-{
-  this->assign(comm);
-}
-
-inline Communicator::~Communicator ()
-{
-  this->clear();
-}
-
-#ifdef LIBMESH_HAVE_MPI
-inline void Communicator::split(int color, int key, Communicator & target) const
-{
-  target.clear();
-  MPI_Comm newcomm;
-  libmesh_call_mpi
-    (MPI_Comm_split(this->get(), color, key, &newcomm));
-
-  target.assign(newcomm);
-  target._I_duped_it = true;
-  target.send_mode(this->send_mode());
-}
-#else
-inline void Communicator::split(int, int, Communicator & target) const
-{
-  target.assign(this->get());
-}
-#endif
-
-inline void Communicator::duplicate(const Communicator & comm)
-{
-  this->duplicate(comm._communicator);
-  this->send_mode(comm.send_mode());
-}
-
-#ifdef LIBMESH_HAVE_MPI
-inline void Communicator::duplicate(const communicator & comm)
-{
-  if (_communicator != MPI_COMM_NULL)
-    {
-      libmesh_call_mpi
-        (MPI_Comm_dup(comm, &_communicator));
-
-      _I_duped_it = true;
-    }
-  this->assign(_communicator);
-}
-#else
-inline void Communicator::duplicate(const communicator &) { }
-#endif
-
-inline void Communicator::clear() {
-#ifdef LIBMESH_HAVE_MPI
-  if (_I_duped_it)
-    {
-      libmesh_assert (_communicator != MPI_COMM_NULL);
-      libmesh_call_mpi
-        (MPI_Comm_free(&_communicator));
-
-      _communicator = MPI_COMM_NULL;
-    }
-  _I_duped_it = false;
-#endif
-}
-
-inline Communicator & Communicator::operator= (const communicator & comm)
-{
-  this->clear();
-  this->assign(comm);
-  return *this;
-}
-
-inline void Communicator::assign(const communicator & comm)
-{
-  _communicator = comm;
-#ifdef LIBMESH_HAVE_MPI
-  if (_communicator != MPI_COMM_NULL)
-    {
-      int i;
-      libmesh_call_mpi
-        (MPI_Comm_size(_communicator, &i));
-
-      libmesh_assert_greater_equal (i, 0);
-      _size = static_cast<unsigned int>(i);
-
-      libmesh_call_mpi
-        (MPI_Comm_rank(_communicator, &i));
-
-      libmesh_assert_greater_equal (i, 0);
-      _rank = static_cast<unsigned int>(i);
-    }
-  else
-    {
-      _rank = 0;
-      _size = 1;
-    }
-#endif
-  _send_mode = DEFAULT;
-}
-
-
-/**
- * Pause execution until all processors reach a certain point.
- */
-#ifdef LIBMESH_HAVE_MPI
-inline void Communicator::barrier () const
-{
-  if (this->size() > 1)
-    {
-      LOG_SCOPE("barrier()", "Parallel");
-      libmesh_call_mpi(MPI_Barrier (this->get()));
-    }
-}
-#else
-inline void Communicator::barrier () const {}
-#endif
-
-
-//-----------------------------------------------------------------------
-// Parallel members
-
-inline
-MessageTag Communicator::get_unique_tag(int tagvalue) const
-{
-  if (used_tag_values.count(tagvalue))
-    {
-      // Get the largest value in the used values, and pick one
-      // larger
-      tagvalue = used_tag_values.rbegin()->first+1;
-      libmesh_assert(!used_tag_values.count(tagvalue));
-    }
-  used_tag_values[tagvalue] = 1;
-
-  // #ifndef NDEBUG
-  //   // Make sure everyone called get_unique_tag and make sure
-  //   // everyone got the same value
-  //   int maxval = tagvalue;
-  //   this->max(maxval);
-  //   libmesh_assert_equal_to (tagvalue, maxval);
-  // #endif
-
-  return MessageTag(tagvalue, this);
-}
-
 #ifdef LIBMESH_HAVE_MPI
 template<>
 inline data_type dataplusint_type<short int>() { return MPI_SHORT_INT; }
