+++ src/libmesh_SOURCES
@@ -260,6 +260,7 @@ libmesh_SOURCES =  \
         src/numerics/trilinos_preconditioner.C \
         src/numerics/type_tensor.C \
         src/numerics/type_vector.C \
+        src/parallel/communicator.C \
         src/parallel/message_tag.C \
         src/parallel/parallel_bin_sorter.C \
         src/parallel/parallel_elem.C \
